[Feat] 키워드로 챌린지를 가져오는 함수 작성
